optimization of TypedArray::Set ----------------------------------------------- https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5S7HF?from=project-issue Add judgment while src is JSStableArray. Change the sequence of type judgment. ----------------------------------------------

+++ ecmascript/base/typed_array_helper-inl.h
@@ -163,9 +163,9 @@ uint32_t TypedArrayHelper::GetSizeFromType(const DataViewType arrayType)
     } else if (arrayType == DataViewType::INT16 ||
                arrayType == DataViewType::UINT16) {
         elementSize = ElementSize::TWO;
-    } else if (arrayType == DataViewType::INT32 ||
+    } else if (arrayType == DataViewType::FLOAT32 ||
                arrayType == DataViewType::UINT32 ||
-               arrayType == DataViewType::FLOAT32) {
+               arrayType == DataViewType::INT32) {
         elementSize = ElementSize::FOUR;
     } else {
         elementSize = ElementSize::EIGHT;
